EAMxx: move scorpio-interface library into its own folder
