+++ components/eamxx/src/physics/mam/mam_aerosol_optics_read_tables.hpp
@@ -5,7 +5,7 @@
 #include "share/manager/field_manager.hpp"
 #include "share/grid/abstract_grid.hpp"
 #include "share/manager/grids_manager.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/io/scorpio_input.hpp"
 
 #include <ekat_parameter_list.hpp>
