+++ components/eamxx/src/physics/mam/readfiles/find_season_index_utils.hpp
@@ -3,7 +3,7 @@
 
 #include <mam4xx/mam4.hpp>
 
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/io/scorpio_input.hpp"
 
 namespace scream::mam_coupling {
