+++ components/eamxx/src/physics/mam/readfiles/marine_organics_impl.hpp
@@ -3,7 +3,7 @@
 
 #include "share/remap/identity_remapper.hpp"
 #include "share/remap/refining_remapper_p2p.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/util/eamxx_timing.hpp"
 
 #include <ekat_team_policy_utils.hpp>
