+++ components/eamxx/src/physics/ml_correction/eamxx_ml_correction_process_interface.hpp
@@ -12,7 +12,7 @@
 #include "share/io/eamxx_output_manager.hpp"
 #include "share/io/scorpio_output.hpp"
 #include "share/io/scorpio_input.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/grid/mesh_free_grids_manager.hpp"
 #include "share/grid/point_grid.hpp"
 #include "share/util/eamxx_time_stamp.hpp"
