+++ components/eamxx/src/physics/rrtmgp/eamxx_rrtmgp_process_interface.cpp
@@ -4,7 +4,7 @@
 #include "physics/rrtmgp/shr_orb_mod_c2f.hpp"
 #include "physics/share/eamxx_trcmix.hpp"
 
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/algorithm/eamxx_fv_phys_rrtmgp_active_gases_workaround.hpp"
 #include "share/property_checks/field_within_interval_check.hpp"
 #include "share/physics/eamxx_common_physics_functions.hpp"
