+++ components/eamxx/src/physics/spa/eamxx_spa_process_interface.cpp
@@ -1,7 +1,7 @@
 #include "eamxx_spa_process_interface.hpp"
 
 #include "share/algorithm/eamxx_data_interpolation.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/property_checks/field_within_interval_check.hpp"
 
 #include <ekat_assert.hpp>
