+++ components/eamxx/src/share/algorithm/eamxx_data_interpolation.cpp
@@ -5,7 +5,7 @@
 #include "share/remap/refining_remapper_p2p.hpp"
 #include "share/remap/iop_remapper.hpp"
 #include "share/grid/point_grid.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/io/scorpio_input.hpp"
 #include "share/io/eamxx_io_utils.hpp"
 #include "share/util/eamxx_universal_constants.hpp"
