+++ components/eamxx/src/share/algorithm/eamxx_time_interpolation.cpp
@@ -1,5 +1,5 @@
 #include "share/algorithm/eamxx_time_interpolation.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/io/eamxx_io_utils.hpp"
 
 namespace scream{
