+++ components/eamxx/src/share/atm_process/IOPDataManager.cpp
@@ -1,7 +1,7 @@
 #include "share/grid/point_grid.hpp"
 #include "share/remap/iop_remapper.hpp"
 #include "share/io/scorpio_input.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/atm_process/IOPDataManager.hpp"
 
 #include <ekat_assert.hpp>
