+++ components/eamxx/src/share/io/eamxx_io_utils.cpp
@@ -1,6 +1,6 @@
 #include "share/io/eamxx_io_utils.hpp"
 
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/manager/library_grids_manager.hpp"
 #include "share/util/eamxx_utils.hpp"
 #include "share/core/eamxx_config.hpp"
