+++ components/eamxx/src/share/io/eamxx_output_manager.cpp
@@ -1,6 +1,6 @@
 #include "eamxx_output_manager.hpp"
 
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/io/scorpio_input.hpp"
 #include "share/util/eamxx_timing.hpp"
 #include "share/core/eamxx_config.hpp"
