+++ components/eamxx/src/share/io/scorpio_output.hpp
@@ -6,7 +6,7 @@
 #include "share/manager/grids_manager.hpp"
 #include "share/grid/abstract_grid.hpp"
 #include "share/io/eamxx_io_utils.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/util/eamxx_time_stamp.hpp"
 #include "share/util/eamxx_utils.hpp"
