+++ components/eamxx/src/share/io/tests/io_alias.cpp
@@ -7,7 +7,7 @@
 #include "share/manager/mesh_free_grids_manager.hpp"
 #include "share/io/eamxx_io_utils.hpp"
 #include "share/io/eamxx_output_manager.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/io/scorpio_input.hpp"
 #include "share/util/eamxx_time_stamp.hpp"
 #include "diagnostics/register_diagnostics.hpp"
