+++ components/eamxx/src/share/io/tests/io_horiz_sampling.cpp
@@ -5,7 +5,7 @@
 
 #include "share/io/eamxx_output_manager.hpp"
 #include "share/io/scorpio_input.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 #include "share/field/field_utils.hpp"
 #include "eamxx_setup_random_test.hpp"
