+++ components/eamxx/src/share/io/tests/io_se_grid.cpp
@@ -2,7 +2,7 @@
 
 #include "share/io/eamxx_output_manager.hpp"
 #include "share/io/scorpio_input.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 
 #include "share/manager/mesh_free_grids_manager.hpp"
