+++ components/eamxx/src/share/io/tests/output_restart.cpp
@@ -3,7 +3,7 @@
 #include "share/io/eamxx_output_manager.hpp"
 #include "share/io/scorpio_output.hpp"
 #include "share/io/scorpio_input.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 
 #include "share/manager/mesh_free_grids_manager.hpp"
 #include "share/grid/point_grid.hpp"
