+++ components/eamxx/src/share/remap/horiz_interp_remapper_data.cpp
@@ -2,7 +2,7 @@
 
 #include "share/grid/point_grid.hpp"
 #include "share/grid/grid_import_export.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 
 #include <numeric>
