+++ components/eamxx/src/share/scorpio_interface/eamxx_scorpio_types.cpp
