+++ components/eamxx/src/share/scorpio_interface/tests/scorpio_interface_tests.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch.hpp>
 
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 
 namespace scream {
