+++ components/eamxx/tests/multi-process/physics_only/shoc_p3_nudging/create_vert_remap_and_weights.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch.hpp>
 #include "share/io/eamxx_output_manager.hpp"
-#include "share/io/eamxx_scorpio_interface.hpp"
+#include "share/scorpio_interface/eamxx_scorpio_interface.hpp"
 
 namespace {
